出品者以外が削除できないように記述

+++ app/controllers/items_controller.rb
@@ -38,6 +38,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    redirect_to root_path unless current_user.id == item.user_id
     item.destroy
     redirect_to root_path
   end
